+++ js/previews.js
@@ -1,5 +1,6 @@
 import './imagefull.js';
-import {fillFullImage, createComments, openFullImage, closeBigPicture} from './imagefull.js';
+import {fillFullImage, createComments} from './imagefull.js';
+import {isEscapeKey} from './util.js';
 
 const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
 const picturesBox = document.querySelector('.pictures');
@@ -21,7 +22,7 @@ const createPreviews = (pictures) => {
     const fullPicture = fillFullImage(picture); //создана полная версия картинки
     const comments = createComments(picture); //созданы комментарии
     const main = document.querySelector('main');
-    const fullPictureCloseElement = fullPicture.querySelector('#picture-cancel');
+    const fullImageCloseElement = fullPicture.querySelector('#picture-cancel');
 
     if (picture.comments.length > 0) {
       const commentsLoader = fullPicture.querySelector('.social__comments-loader');
@@ -38,9 +39,33 @@ const createPreviews = (pictures) => {
     });
 
     //закрытие полноразмерной картинки
-    fullPictureCloseElement.addEventListener('click', () => {
-      closeBigPicture(fullPicture);
+    fullImageCloseElement.addEventListener('click', () => {
+      closeFullImage(fullPicture);
     });
+
+    //Закрытие полной версии картинки при нажатии Esc
+    const onDocumentKeydown = (evt) => {
+      if (isEscapeKey(evt)) {
+        evt.preventDefault();
+        closeFullImage(fullPicture);
+      }
+    };
+
+    //Открытие полной версии картинки
+    function openFullImage () {
+      fullPicture.classList.remove('hidden');
+      document.body.classList.add('modal-open');
+
+      document.addEventListener('keydown', onDocumentKeydown);
+    }
+
+    //Закрытие полной версии картинки
+    function closeFullImage () {
+      fullPicture.classList.add('hidden');
+      document.body.classList.remove('modal-open');
+
+      document.removeEventListener('keydown', onDocumentKeydown);
+    }
   });
   picturesBox.append(previewsListFragment);
 };
